Change #1
While working on this section of my code, I added an image, I changed
the background, and I also added another edited image. In order to
upload an image into processing, I first had to upload it into the
sketch folder, which then allowed me to name it in processing. I also
wanted my game to be able to move the character with the mouse, so I
added a code that said:
image (donald, mouseX, mouseY);
This is essentially saying, that I want the mouse to be able to move
the character around anywhere on the screen; whether it is up, down or
side to side.