Change #3
The changes I made are obvious, as most of it was to create a moving
ellipse across the screen. In order to do that, I first needed to add
the speed, the randomised direction (xposition and position), and  the
int rad to indicate the radius of which the ellipse is moving at on the
screen. In the void setup area, I wanted the x and y position to move
in different directions each time. Therefore, I divided the width/2 and
the height/2 so the ellipse isn’t moving in one way only. As well as
that, further down the code, I researched about how to get the ellipse
to constantly move in the x and y direction, though different each
time. Which is why I added that if the x position is bigger than the
width minus the radius, the ellipse would move in a direction that is
bigger than the radius of the ellipse/screen.